+++ packages/trpc-playground/handlers/aws-lambda.d.ts
@@ -1,0 +1,1 @@
+export * from '../dist/handlers/aws-lambda'