+++ packages/trpc-playground/src/handlers/aws-lambda.ts
@@ -1,0 +1,4 @@
+import { getAwsLambdaAdapter } from 'uttp/adapters/aws-lambda'
+import { handler } from '../handler'
+
+export const awsLambdaHandler = getAwsLambdaAdapter(handler)