+++ packages/trpc-playground/src/handlers/express.ts
@@ -1,12 +1,4 @@
-import { TrpcPlaygroundConfig } from '@trpc-playground/types'
-import { Handler } from 'express'
-import { getCommonHandlerReqData } from './common'
-import { nodeHttpHandler } from './node-http'
+import { getExpressAdapter } from 'uttp/adapters/express'
+import { handler } from '../handler'
 
-export const expressHandler = async (config: TrpcPlaygroundConfig): Promise<Handler> => {
-  const common = await getCommonHandlerReqData(config)
-
-  return async (req, res) => {
-    await nodeHttpHandler({ req, res, common })
-  }
-}
+export const expressHandler = getExpressAdapter(handler)