+++ packages/trpc-playground/src/handlers/fetch.ts
@@ -1,0 +1,4 @@
+import { getFetchAdapter } from 'uttp/adapters/fetch'
+import { handler } from '../handler'
+
+export const fetchHandler = getFetchAdapter(handler)